Initial string-based exclude mechanism
The implementation takes every path in a given path and checks if it is
mentioned in the exclude vector. If that isn't the case, push it onto a
new vector which only contains non-excluded paths. At the end, print all
non-excluded files.

INFO: Waites on the parser implementation and the split between
`exclude_strings` and `exclude_regex`