Implement moving files to backup location
The current implementations can handle files and directories. If a file
already exists in the backup folder, it checks the last modified time to
determine if it should be overridden.

The implementations is now split into two files: path_filter which
concerns itself with filtering paths, as the name suggests. The other
file is named backup, where the main logic of the backup process happens
(comparing timestamps, moving files, moving directories recursively)