Implement chunking (#6)
Chunking is now implemented. The program initializes a repository structure. The structure contains a directory "chunks" which holds all chunks of files backed up. Additionally, there is an "index.index" file which stores the files metadata and which chunks it is comprised of.

Backup now has two modes in which it can function. The first mode is `backup` where the first argument is the data to back up and the second is a repository (which can exist). 
The other mode is the restore mode, where files get pieced back together from chunks and written to the second argument.

Furthermore, there are multiple style fixes and code quality improvements